#953 TkServletApp renamed and put inside SrvTakeTest
