--- src/test/java/org/takes/servlet/TkServletApp.java
@@ -1,66 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2014-2019 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package org.takes.servlet;
-
-import java.io.IOException;
-import javax.servlet.ServletContext;
-import org.cactoos.text.FormattedText;
-import org.cactoos.text.UncheckedText;
-import org.takes.Request;
-import org.takes.Response;
-import org.takes.Take;
-import org.takes.rs.RsText;
-
-/**
- * Fake TkServletApp (for {@link SrvTake} test only).
- *
- * @since 1.16
- */
-public final class TkServletApp implements Take {
-    /**
-     * ServletContext.
-     */
-    private final ServletContext context;
-
-    /**
-     * Ctor.
-     * @param ctx A ServletContext
-     */
-    public TkServletApp(final ServletContext ctx) {
-        this.context = ctx;
-    }
-
-    @Override
-    public Response act(final Request req) throws IOException {
-        return new RsText(
-            new UncheckedText(
-                new FormattedText(
-                    "Hello servlet! Using [%s] ServletContext.",
-                    this.context.getServletContextName()
-                )
-            ).asString()
-        );
-    }
-}
